+++ superset-frontend/src/dashboard/components/nativeFilters/state.ts
@@ -1,0 +1,48 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { useSelector } from 'react-redux';
+import { useMemo } from 'react';
+import { Filter } from './types';
+import { FilterConfiguration } from './FilterConfigModal/types';
+
+const defaultFilterConfiguration: Filter[] = [];
+
+export function useFilterConfiguration() {
+  return useSelector<any, FilterConfiguration>(
+    state =>
+      state.dashboardInfo?.metadata?.filter_configuration ||
+      defaultFilterConfiguration,
+  );
+}
+
+/**
+ * returns the dashboard's filter configuration,
+ * converted into a map of id -> filter
+ */
+export function useFilterConfigMap() {
+  const filterConfig = useFilterConfiguration();
+  return useMemo(
+    () =>
+      filterConfig.reduce((acc: Record<string, Filter>, filter: Filter) => {
+        acc[filter.id] = filter;
+        return acc;
+      }, {} as Record<string, Filter>),
+    [filterConfig],
+  );
+}
