+++ superset-frontend/src/dashboard/reducers/types.ts
@@ -73,7 +73,7 @@ export type LayoutItem = {
 };
 
 /** Current state of the filter, stored in `nativeFilters` in redux */
-export type FilterState = {
+export type NativeFilterState = {
   id: string; // ties this filter state to the config object
   extraFormData?: ExtraFormData;
   currentState?: CurrentFilterState;
@@ -84,6 +84,6 @@ export type NativeFiltersState = {
     [filterId: string]: Filter;
   };
   filtersState: {
-    [filterId: string]: FilterState;
+    [filterId: string]: NativeFilterState;
   };
 };
